Added difficulty selector and a hard difficulty.
Added a feature that adds a start text that allows the player to select a difficulty in the beginning. There are two difficulties, Regular or Hard. If R is enter, it will run the game as it was implemented before. If H is entered, an item will be randomly placed into one of the rooms (excluding the exist room) that the player will have to find in order to get to the exist.